xdiff: upgrade to git's latest xdiff version
Update to the latest version of xdiff from upstream git.  Remove
attempts within xdiff to move into `size_t` types for the length of
buffers.  Instead, ensure that we never pass `LONG_MAX` into xdiff.

Also, examine variables and their ranges, to allow us to remove some
overflow checking.

+++ src/xdiff/xpatience.c
@@ -78,7 +78,7 @@ struct hashmap {
 
 static int is_anchor(xpparam_t const *xpp, const char *line)
 {
-	unsigned long i;
+	size_t i;
 	for (i = 0; i < xpp->anchors_nr; i++) {
 		if (!strncmp(line, xpp->anchors[i], strlen(xpp->anchors[i])))
 			return 1;
